chore: move social data
